# docs/swagger.yaml
...
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.GetReviewingTypeResp:
    properties:
      label:
        type: string
      name:
        type: string
      todo_amount:
        type: integer
    type: object
  schema.GetRevisionResp:
    properties:
      content: {}
      create_at:
        type: integer
      id:
        type: string
      object_id:
        type: string
      reason:
        type: string
      status:
        type: integer
      title:
        type: string
      url_title:
        type: string
      use_id:
        type: string
      user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
    type: object
...
  schema.GetSiteLegalInfoResp:
    properties:
      privacy_policy_original_text:
        type: string
      privacy_policy_parsed_text:
        type: string
      terms_of_service_original_text:
        type: string
      terms_of_service_parsed_text:
        type: string
    type: object
...
  schema.GetUnreviewedRevisionResp:
    properties:
      info:
        $ref: '#/definitions/schema.UnreviewedRevisionInfoInfo'
      type:
        type: string
      unreviewed_info:
        $ref: '#/definitions/schema.GetRevisionResp'
    type: object
...
  schema.RevisionAuditReq:
    properties:
      id:
        description: object id
        type: string
      operation:
        description: approve or reject
        type: string
    required:
    - id
    - operation
    type: object
...
  schema.TagResp:
    properties:
      display_name:
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
    type: object
...
  schema.UnreviewedRevisionInfoInfo:
    properties:
      answer_accepted:
        type: boolean
      answer_count:
        type: integer
      answer_id:
        type: string
      comment_id:
        type: string
      content:
        type: string
      created_at:
        type: integer
      html:
        type: string
      object_creator_user_id:
        type: string
      object_id:
        type: string
      object_type:
        type: string
      question_id:
        type: string
      show_status:
        type: integer
      status:
        type: integer
      tags:
        items:
          $ref: '#/definitions/schema.TagResp'
        type: array
      title:
        type: string
      url_title:
        type: string
    type: object
...
  schema.UserBasicInfo:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      id:
        type: string
      language:
        type: string
      location:
        type: string
      rank:
        type: integer
      status:
        type: string
      username:
        type: string
      website:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/reviewing/type:
    get:
      description: get reviewing type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.GetReviewingTypeResp'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: get reviewing type
      tags:
      - Revision
  /answer/api/v1/revisions:
    get:
      description: get revision list
      parameters:
      - description: object id
        in: query
        name: object_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.GetRevisionResp'
                  type: array
              type: object
      summary: get revision list
      tags:
      - Revision
  /answer/api/v1/revisions/audit:
    put:
      description: revision audit operation:approve or reject
      parameters:
      - description: audit
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RevisionAuditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: revision audit
      tags:
      - Revision
  /answer/api/v1/revisions/edit/check:
    get:
      consumes:
      - application/json
      description: check can update revision
      parameters:
      - default: string
        description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: check can update revision
      tags:
      - Revision
  /answer/api/v1/revisions/unreviewed:
    get:
      description: get unreviewed revision list
      parameters:
      - description: page id
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetUnreviewedRevisionResp'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: get unreviewed revision list
      tags:
      - Revision
...
  /answer/api/v1/siteinfo/legal:
    get:
      description: get site legal info
      parameters:
      - description: legal information type
        enum:
        - tos
        - privacy
        in: query
        name: info_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetSiteLegalInfoResp'
              type: object
      summary: get site legal info
      tags:
      - site
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"